Refactor diag to avoid circular dependencies
