+++ include/tvm/ir/transform.h
@@ -56,7 +56,7 @@
 #ifndef TVM_IR_TRANSFORM_H_
 #define TVM_IR_TRANSFORM_H_
 
-#include <tvm/ir/diagnostic.h>
+#include <tvm/ir/diagnostic_context.h>
 #include <tvm/ir/error.h>
 #include <tvm/ir/module.h>
 #include <tvm/node/container.h>
