+++ include/tvm/ir/type_relation.h
@@ -26,6 +26,7 @@
 
 #include <tvm/ir/attrs.h>
 #include <tvm/ir/diagnostic.h>
+#include <tvm/ir/diagnostic_context.h>
 #include <tvm/ir/env_func.h>
 #include <tvm/ir/module.h>
 #include <tvm/ir/type.h>
